move RecursionGuard, optimise recursion check
